requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.3.0
python-docx>=1.1.0
PyPDF2>=3.0.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0 